WEBAPP-208: Fix sending of feedback on opening
